importer-service/requirements.txt
fastapi
uvicorn
requests
redis
google-auth-oauthlib
google-api-python-client
